🚚 move recoverable-error-boundary back to hamr
