--- packages/hamr/react-error-boundary/src/index.ts
@@ -1,1 +1,0 @@
-export * from "~/packages/atom.io/react-devtools/src/error-boundary"
